feat: (WIP) start working on path migration
The rewrite features a complete removal of the internal command tree. This commit marks an initial
step towards that migration. Some old, removed classes are still not cleaned up in the relevant
places and the new path system is not fully implemented yet and not at all hooked up correctly.