feat: split into -common and -paper modules
A lot happened here, but the biggest change by far is the separation away from the horrible compiler API, which was causing me problems and headaches
and abstracting it away into my own (rather big) abstraction, providing everything that I need. This has a few advantages, including (but not limited to):
1. I can finally start writing up units tests, as thanks to the abstraction, I can mock these types now.
2. I can get rid of the horrendously complex Utils class which was hiding away all of the 'ugly' logic.
3. This has furthermore allowed me to pass around the ProcessingEnvironment everywhere I need to mess with Elements/Types (which was basically everwhere), since it lives in the SourceX implementation files now.

The processor-X and anotations-X modules compile now. I haven't tested whether this has borked the annotation processor as a whole yet, and also the Gradle setup might be a bit cursed atm. This commit simply marks the initial separation.